Shrink the size of `Amode`
Reduces the size of `KnownOffset` so `Option<KnownOffset>` doesn't have
such a high alignment or size, greatly shrinking the size of the
instruction and prevents the changes to `Div` from increasing the size
of the overall `Inst` enum. In fact it now shrinks!

+++ cranelift/assembler-x64/src/mem.rs
@@ -125,7 +125,7 @@ impl AmodeOffsetPlusKnownOffset {
     #[must_use]
     pub fn value(&self, offsets: &impl KnownOffsetTable) -> i32 {
         let known_offset = match self.offset {
-            Some(offset) => offsets[offset],
+            Some(offset) => offsets[usize::from(offset)],
             None => 0,
         };
         known_offset
