+++ cranelift/codegen/src/isa/x64/inst/emit.rs
@@ -4101,9 +4101,9 @@ pub(crate) fn emit(
             // assembler; due to when Cranelift determines these offsets, this
             // happens quite late (i.e., here during emission).
             let frame = state.frame_layout();
-            known_offsets[external::offsets::KEY_INCOMING_ARG] =
+            known_offsets[usize::from(external::offsets::KEY_INCOMING_ARG)] =
                 i32::try_from(frame.tail_args_size + frame.setup_area_size).unwrap();
-            known_offsets[external::offsets::KEY_SLOT_OFFSET] =
+            known_offsets[usize::from(external::offsets::KEY_SLOT_OFFSET)] =
                 i32::try_from(frame.outgoing_args_size).unwrap();
             inst.encode(sink, &known_offsets);
         }
